+++ fiktion-core/src/commonMain/kotlin/dev/s7a/fiktion/GenerateValue.kt
@@ -85,7 +85,7 @@ internal fun generateAutomaticValue(
         )
     }
 
-    throw CannotGenerateException(missingGenerationMessage(request.type))
+    throw CannotGenerateException(missingGenerationMessage(request = request, config = config))
 }
 
 /**